Entities package
This is one of the fundamental packages to allow the game to work. This creates entities that the modules can be rendered through and a basis for all models. This is also where the player is created and how they move through the world as well as the "camera" that we see the game through.